+++ try.go
@@ -55,18 +55,17 @@ func GenerateRequest(testCase *TestCase) (*http.Request, error) {
 	} else {
 		req, err = http.NewRequest(testCase.Request.Method, testCase.Request.Url, bytes.NewBuffer(reqJson))
 	}
-
 	if err != nil {
 		return nil, err
 	}
 
-	// Should always request JSON
+	// Set some default headers for simplicity. These can be overwritten further down on each test case if required.
 	if testCase.RequestContentType != "" {
 		req.Header.Set(echo.HeaderContentType, testCase.RequestContentType)
 	} else {
 		req.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
 	}
-
+	req.Header.Set("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/107.0.0.0 Safari/537.36")
 	req.Header.Set(echo.HeaderXRealIP, "127.0.0.0")
 
 	// Add cookies in if present