+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [staging]
+### Added
+- Added a default user agent in, which cna be overriten by the headers in a test case
 
 ## [1.1.1] - 2026-01-13 (96.2%)
 ### Fixed